+++ src/biodwh2-interpro/src/main/java/de/unibi/agbi/biodwh2/interpro/model/DelRef.java
@@ -1,0 +1,8 @@
+package de.unibi.agbi.biodwh2.interpro.model;
+
+import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
+
+public class DelRef {
+    @JacksonXmlProperty(isAttribute = true)
+    public String id;
+}