+++ src/biodwh2-interpro/src/main/java/de/unibi/agbi/biodwh2/interpro/model/Release.java
@@ -1,0 +1,12 @@
+package de.unibi.agbi.biodwh2.interpro.model;
+
+import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlElementWrapper;
+import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
+
+import java.util.List;
+
+public class Release {
+    @JacksonXmlElementWrapper(useWrapping = false)
+    @JacksonXmlProperty(localName = "dbinfo")
+    public List<DBInfo> dbInfo;
+}